requirements.txt
#Gerenciamento de Dados
pandas
SQLAlchemy
pymysql
python-dotenv

#Modelagem
scikit-learn
statsmodels
prophet
xgboost

#Serviço de API
fastapi
uvicorn

pyarrow